Try to improve performance

+++ ising.c
@@ -46,7 +46,7 @@ void update(const float temp, int8_t grid[L][L]) {
 
 
 double calculate(int8_t grid[L][L], int *M_max) {
-  int8_t E = 0;
+  int E = 0;
   for (unsigned int i = 0; i < L; ++i) {
     for (unsigned int j = 0; j < L; ++j) {
       int8_t spin = grid[i][j];
